+++ engines/ep/src/ep_vb.h
@@ -136,7 +136,7 @@ public:
      * BgFetchers assigned to this vBucket, the distributionKey is used to
      * select a BgFetcher from that set.
      */
-    BgFetcher& getBgFetcher(uint32_t distributionKey);
+    BgFetcher& getBgFetcher(uint64_t distributionKey);
     Flusher* getFlusher() override;
 
     UniqueDCPBackfillPtr createDCPBackfill(EventuallyPersistentEngine& e,
